PR76 t3 refactor
