+++ .logs/t3_plan.jsonl
@@ -1,1 +1,1 @@
-{"turn": "demo-1", "agent": "AgentA", "policy_backend": "rulebased", "backend": "rulebased", "ops_counts": {"Speak": 1, "RequestRetrieve": 1}, "requested_retrieve": true, "rag_used": true, "reflection": false, "ms_deliberate": 0.058, "ms_rag": 0.086, "now": "2025-10-02T15:08:15.070224+00:00"}
+{"turn": "demo-1", "agent": "AgentA", "policy_backend": "rulebased", "backend": "rulebased", "ops_counts": {"Speak": 1, "RequestRetrieve": 1}, "requested_retrieve": true, "rag_used": true, "reflection": false, "ms_deliberate": 0.086, "ms_rag": 0.075, "now": "2025-10-03T01:02:02.077688+00:00"}
