+++ .logs/turn.jsonl
@@ -1,1 +1,1 @@
-{"turn": "demo-1", "agent": "AgentA", "durations_ms": {"t1": 0.145, "t2": 0.285, "t4": 0.014, "apply": 0.144, "total": 1.892}, "t1": {"pops": 5, "iters": 1, "graphs_touched": 1}, "t2": {"k_returned": 0, "k_used": 0, "cache_hit": false}, "t4": {"approved": 0, "rejected": 0}, "now": "2025-10-02T15:08:15.070224+00:00"}
+{"turn": "demo-1", "agent": "AgentA", "durations_ms": {"t1": 0.178, "t2": 0.237, "t4": 0.013, "apply": 0.15, "total": 2.06}, "t1": {"pops": 5, "iters": 1, "graphs_touched": 1}, "t2": {"k_returned": 0, "k_used": 0, "cache_hit": false}, "t4": {"approved": 0, "rejected": 0}, "now": "2025-10-03T01:02:02.077688+00:00"}
