+++ python
@@ -1,0 +1,2 @@
+#!/bin/sh
+exec python3 "$@"
