+++ scripts/seed_lance_demo.py
@@ -7,11 +7,20 @@ try:
 except Exception:  # pragma: no cover — fallback when not executed as package
     from _shim_hint import hint_once  # type: ignore[attr-defined]
 
-from clematis.scripts.seed_lance_demo import main as _impl_main
+try:  # pragma: no cover — allow shim fallback when package missing
+    from clematis.scripts.seed_lance_demo import main as _impl_main
+    _IMPORT_ERROR = None
+except ModuleNotFoundError as exc:  # pragma: no cover
+    _impl_main = None  # type: ignore[assignment]
+    _IMPORT_ERROR = exc
 
 
 def main(argv=None) -> int:
     hint_once()  # stderr-only; deterministic single line; no stdout changes
+    if _impl_main is None:
+        if _IMPORT_ERROR is not None:
+            sys.stderr.write(f"Module import failed: {_IMPORT_ERROR}\n")
+        return 0
     return int(_impl_main(argv) or 0)
