+++ tests/conftest.py
@@ -6,6 +6,20 @@ ROOT = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
 if ROOT not in sys.path:
     sys.path.insert(0, ROOT)
 
+# Ensure the interpreter's bin directory is on PATH so subprocess calls to
+# `python` resolve even on environments where only `python3` is registered.
+_exe_dir = os.path.dirname(sys.executable or "")
+paths = os.environ.get("PATH", "")
+segments = [p for p in paths.split(os.pathsep) if p]
+
+if _exe_dir and _exe_dir not in segments:
+    segments.insert(0, _exe_dir)
+
+if ROOT not in segments:
+    segments.insert(0, ROOT)
+
+os.environ["PATH"] = os.pathsep.join(segments)
+
 # CI/offline guard: ban outbound sockets when CLEMATIS_NETWORK_BAN=1
 import socket
 import pytest
